+++ save-orchestrator/src/main/kotlin/com/saveourtool/save/orchestrator/docker/DockerPersistentVolumeService.kt
@@ -1,25 +1,24 @@
 package com.saveourtool.save.orchestrator.docker
 
-import com.saveourtool.save.orchestrator.config.ConfigProperties
-import com.saveourtool.save.orchestrator.service.PersistentVolumeId
-import com.saveourtool.save.orchestrator.service.PersistentVolumeService
-
 import com.github.dockerjava.api.DockerClient
 import com.github.dockerjava.api.command.PullImageResultCallback
 import com.github.dockerjava.api.model.HostConfig
 import com.github.dockerjava.api.model.Mount
 import com.github.dockerjava.api.model.MountType
+import com.saveourtool.save.orchestrator.config.ConfigProperties
+import com.saveourtool.save.orchestrator.service.PersistentVolumeService
 import org.springframework.context.annotation.Profile
 import org.springframework.stereotype.Component
 import reactor.core.publisher.Flux
-
 import java.nio.file.Path
-import java.util.UUID
-
+import java.util.*
 import kotlin.io.path.absolutePathString
 import kotlin.time.Duration.Companion.seconds
 import kotlin.time.toJavaDuration
 
+/**
+ * Implementation of [PersistentVolumeService] that creates [Docker Volumes](https://docs.docker.com/storage/volumes/)
+ */
 @Component
 @Profile("!kubernetes")
 class DockerPersistentVolumeService(
@@ -34,11 +33,7 @@ class DockerPersistentVolumeService(
         val createVolumeResponse = dockerClient.createVolumeCmd()
             .withName("save-execution-vol-${UUID.randomUUID()}")
             .exec()
-        dockerClient.pullImageCmd("alpine")
-            .withRegistry(configProperties.docker.registry)
-            .withTag("latest")
-            .exec(PullImageResultCallback())
-            .awaitCompletion()
+        blockingPullImage("alpine", "latest")
         val createContainerResponse = dockerClient.createContainerCmd("alpine:latest")
             .withHostConfig(
                 HostConfig()
@@ -62,27 +57,38 @@ class DockerPersistentVolumeService(
             )
             .exec()
         val dataCopyingContainerId = createContainerResponse.id
+
         dockerClient.startContainerCmd(dataCopyingContainerId)
             .exec()
-        val copyingTimeout = 200.seconds
-        val checkInterval = 10.seconds
-        Flux.interval(checkInterval.toJavaDuration())
-            .take((copyingTimeout / checkInterval).toLong())
-            .takeWhile {
-                val inspectContainerResponse = dockerClient.inspectContainerCmd(dataCopyingContainerId).exec()
-                inspectContainerResponse.state.status == "running"
-            }
-            .blockLast()
+        waitForCompletionWithTimeout(dataCopyingContainerId)
 
         dockerClient.removeContainerCmd(dataCopyingContainerId)
 
         return DockerPvId(createVolumeResponse.name)
     }
+
+    private fun blockingPullImage(
+        repository: String,
+        tag: String
+    ) = dockerClient.pullImageCmd(repository)
+        .withRegistry(configProperties.docker.registry)
+        .withTag(tag)
+        .exec(PullImageResultCallback())
+        .awaitCompletion()
+
+    private fun waitForCompletionWithTimeout(containerId: String) {
+        val copyingTimeout = 200.seconds
+        val checkInterval = 10.seconds
+        Flux.interval(checkInterval.toJavaDuration())
+            .take((copyingTimeout / checkInterval).toLong())
+            .takeUntil {
+                val inspectContainerResponse = dockerClient.inspectContainerCmd(containerId).exec()
+                inspectContainerResponse.state.status == "running"
+            }
+            .blockLast()
+            .let {
+                requireNotNull(it) { "Container $containerId still running after $copyingTimeout" }
+            }
+    }
 }
 
-/**
- * @property volumeName name of the Docker volume
- */
-data class DockerPvId(
-    val volumeName: String,
-) : PersistentVolumeId
