+++ save-orchestrator/src/main/kotlin/com/saveourtool/save/orchestrator/kubernetes/KubernetesPvId.kt
@@ -1,0 +1,10 @@
+package com.saveourtool.save.orchestrator.kubernetes
+
+import com.saveourtool.save.orchestrator.service.PersistentVolumeId
+
+/**
+ * @property id ID of the Kubernetes PV
+ */
+data class KubernetesPvId(
+    val id: String
+) : PersistentVolumeId
