+++ save-orchestrator/src/main/kotlin/com/saveourtool/save/orchestrator/service/PersistentVolumeService.kt
@@ -1,7 +1,14 @@
+/**
+ * Interact with persistent volumes of a certain storage system.
+ */
+
 package com.saveourtool.save.orchestrator.service
 
 import java.nio.file.Path
 
+/**
+ * Interface that lists method to interact with persistent volumes of a certain storage system.
+ */
 interface PersistentVolumeService {
     /**
      * @param resources list of Paths that should be copied into a persistent volume
@@ -10,4 +17,7 @@ interface PersistentVolumeService {
     fun createFromResources(resources: Collection<Path>): PersistentVolumeId
 }
 
+/**
+ * Identifier of a persistent volume
+ */
 interface PersistentVolumeId
