+++ save-orchestrator/src/main/kotlin/com/saveourtool/save/orchestrator/utils/LoggingContext.kt
@@ -5,3 +5,5 @@ import org.slf4j.Logger
 interface LoggingContext {
     val logger: Logger
 }
+
+class LoggingContextImpl(override val logger: Logger): LoggingContext
